Move kit assocation from DetailedLibrary to Library
JIRA: GLT-471
