+++ miso-dto/src/main/java/uk/ac/bbsrc/tgac/miso/dto/Dtos.java
@@ -898,9 +898,6 @@ public class Dtos {
 
   private static DetailedLibraryDto asDetailedLibraryDto(DetailedLibrary from) {
     DetailedLibraryDto dto = new DetailedLibraryDto();
-    if (from.getKitDescriptor() != null) {
-      dto.setKitDescriptorId(from.getKitDescriptor().getId());
-    }
     if (from.getLibraryDesign() != null) {
       dto.setLibraryDesignId(from.getLibraryDesign().getId());
     }
@@ -917,11 +914,6 @@ public class Dtos {
   public static DetailedLibrary toDetailedLibrary(DetailedLibraryDto from) {
     if (from == null) return null;
     DetailedLibrary to = new DetailedLibraryImpl();
-    if (from.getKitDescriptorId() != null) {
-      KitDescriptor kitDescriptor = new KitDescriptor();
-      kitDescriptor.setId(from.getKitDescriptorId());
-      to.setKitDescriptor(kitDescriptor);
-    }
     if (from.getLibraryDesignId() != null) {
       LibraryDesign design = new LibraryDesign();
       design.setId(from.getLibraryDesignId());
@@ -1035,6 +1027,9 @@ public class Dtos {
       dto.setPlatformType(from.getPlatformType().getKey());
     }
     dto.setLastModified(formatDateTime(from.getLastModified()));
+    if (from.getKitDescriptor() != null) {
+      dto.setKitDescriptorId(from.getKitDescriptor().getId());
+    }
     if (!from.getIndices().isEmpty()) {
       dto.setIndexFamilyName(from.getIndices().get(0).getFamily().getName());
       for (Index index : from.getIndices()) {
@@ -1130,6 +1125,11 @@ public class Dtos {
       to.setVolume(Double.valueOf(from.getVolume()));
     }
     to.setDnaSize(from.getDnaSize());
+    if (from.getKitDescriptorId() != null) {
+      KitDescriptor kitDescriptor = new KitDescriptor();
+      kitDescriptor.setId(from.getKitDescriptorId());
+      to.setKitDescriptor(kitDescriptor);
+    }
     to.setLocationBarcode(from.getLocationBarcode());
     to.setCreationDate(parseDate(from.getCreationDate()));
